+++ src/xcode/ENA/ENA/Source/Scenes/HealthCertificate/Models/HealthCertificate.swift
@@ -1,0 +1,34 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import OpenCombine
+
+struct HealthCertificate: Codable {
+
+	// MARK: - Internal
+
+	let representations: HealthCertificateRepresentations
+
+	var version: String {
+		healthCertificateResponse.version
+	}
+
+	var name: HealthCertificateName {
+		healthCertificateResponse.name
+	}
+
+	var dateOfBirth: String {
+		healthCertificateResponse.dateOfBirth
+	}
+
+	var vaccinationCertificates: [VaccinationCertificate] {
+		healthCertificateResponse.vaccinationCertificates
+	}
+
+	// MARK: - Private
+
+	private let healthCertificateResponse: HealthCertificateResponse
+
+}
