+++ src/xcode/ENA/ENA/Source/Scenes/HealthCertificate/Models/HealthCertificateRepresentations.swift
@@ -1,0 +1,16 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import OpenCombine
+
+struct HealthCertificateRepresentations: Codable {
+
+	// MARK: - Internal
+
+	let base45: String
+	let cbor: Data
+	let json: Data
+
+}
