+++ src/xcode/ENA/ENA/Source/Scenes/HealthCertificate/Models/ProofCertificate.swift
@@ -1,0 +1,15 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import OpenCombine
+
+struct ProofCertificate: Codable {
+
+	// MARK: - Internal
+
+	let cborRepresentation: Data
+	let expirationDate: Date
+
+}
